% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distill_website.R
\name{rjdistill_website}
\alias{rjdistill_website}
\title{R Markdown site generator for Distill websites}
\usage{
rjdistill_website(input, encoding = getOption("encoding"), ...)
}
\arguments{
\item{input}{Website directory (or the name of a file within the directory).}

\item{encoding}{Ignored. The encoding is always assumed to be UTF-8.}

\item{...}{Currently unused.}
}
\description{
R Markdown site generator for Distill websites
}
